subir modificaciones de migraciones

+++ database/migrations/2021_09_21_072057_create_orders_table.php
@@ -4,6 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
+
 class CreateOrdersTable extends Migration
 {
     /**
@@ -15,7 +16,7 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->integer('id_user');
+            $table->unsignedBigInteger('id_user');
             $table->integer('amount');
             $table->string('address');
             $table->string('email');
@@ -27,6 +28,7 @@ class CreateOrdersTable extends Migration
             $table->foreign('id_user')->references('id')->on('users');
 
         });
+          
     }
 
     /**
