+++ database/migrations/2021_09_21_072829_create_products_table.php
@@ -21,7 +21,7 @@ class CreateProductsTable extends Migration
             $table->integer('weight');
             $table->string('description');
             $table->binary('image');
-            $table->integer('id_category');
+            $table->unsignedBigInteger('id_category');
             $table->integer('stock');
             $table->timestamps();
